requirements.txt
click==8.1.3
colorama==0.4.5
Flask==2.2.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
MarkupSafe==2.1.1
numpy==1.23.4
scikit-learn==1.1.2
scipy==1.9.2
sklearn==0.0
threadpoolctl==3.1.0
Werkzeug==2.2.2
tensorflow-cpu==2.11.0
keras==2.11.0
h5py==3.7.0
pandas==1.5.2
pmdarima==2.0.1